feat: add autoFixContext function
