+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-activation",
-  "version": "0.7.4",
+  "version": "0.8.0-beta.1",
   "description": "<KeepAlive /> for React like <keep-alive /> in vue",
   "main": "index.js",
   "scripts": {
