+++ src/core/Bridge/Context/index.js
@@ -1,0 +1,5 @@
+import { fixContext, autoFixContext, createContext } from './fixContext'
+
+export { fixContext, autoFixContext, createContext }
+export { default as ProviderBridge } from './ProviderBridge'
+export { default as ConsumerBridge } from './ConsumerBridge'
