+++ src/index.js
@@ -1,7 +1,7 @@
 import AliveScope from './core/AliveScope'
 import { withActivation, useActivate, useUnactivate } from './core/lifecycles'
 import KeepAlive from './core/KeepAlive'
-import { fixContext, createContext } from './core/Bridge/Context'
+import { fixContext, createContext, autoFixContext } from './core/Bridge/Context'
 import withAliveScope, { useAliveController } from './core/withAliveScope'
 import NodeKey from './core/NodeKey'
 
@@ -11,6 +11,7 @@ export {
   AliveScope,
   withActivation,
   fixContext,
+  autoFixContext,
   useActivate,
   useUnactivate,
   createContext,
